feat: support macOS runners
Currently the wonderland2-setup-action fails on macOS runners because they are not supported by the unfor19/install-aws-cli-action used. The action does not intend to add support: https://github.com/unfor19/install-aws-cli-action/issues/10

However, this step actually should not be necessary on public runners and was only added for internal ones.

Therefore, skip the aws step on macOS and download the correct build.